I kept dfc_rev as long so that it can hold a long time_t from stat.st_mtime, so I forgot to change this pointer type back.  Fix some elog specifiers too.
git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@8307 3f9b6c58-9310-455a-aec6-8b30269ab333

+++ src/ds/DataFormat.c
@@ -564,8 +564,8 @@ dfa_OpenFile (const DataFile *df, int write)
 	/*
 	 * Pointer to the rev in the DF portion of our open file
 	 */
-		int *ofrev = &(ofp->of_df.df_core.dfc_rev);
-	    
+		long *ofrev = &(ofp->of_df.df_core.dfc_rev);
+
                 if (write && ! ofp->of_write)
                 {
 #ifdef DEBUG
@@ -581,9 +581,9 @@ dfa_OpenFile (const DataFile *df, int write)
                  * The latest data file entry has a new revision, so our
                  * open file's tag must be out of date.  Thus sync the file. 
                  */
-                        msg_ELog (EF_DEBUG, "file %s out of sync: %d < %d",
-                                  df->df_fullname, *ofrev,
-				  df->df_core.dfc_rev);
+                        msg_ELog (EF_DEBUG, "file %s out of sync: %ld < %ld",
+                                  df->df_fullname, (long)*ofrev,
+				  (long)df->df_core.dfc_rev);
 #ifdef NCSYNC_FIXED
 			dfa_SyncFile (ofp);
                         *ofrev = df->df_core.dfc_rev;
